% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGAcolon_adenocarcinomaTPM}
\alias{TCGAcolon_adenocarcinomaTPM}
\title{TCGA expression sample data}
\format{
A data frame
}
\source{
\url{https://portal.gdc.cancer.gov}
}
\usage{
data(TCGAcolon_adenocarcinomaTPM)
}
\description{
colon_adenocarcinoma TPM values from TCGA
}
\keyword{datasets}
